requirements.txt
apache-airflow
requests
numpy>=1.22.0
pandas
python-dotenv
sendgrid
apache-airflow-providers-google
apache-airflow-providers-postgres
psycopg2-binary==2.9.10
pytest
pytest-cov
pytz
apache-airflow-providers-vertica
